requirements.txt
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
filelock==3.10.6
huggingface-hub==0.13.4
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.0
numpy==1.24.2
packaging==23.0
pathy==0.10.1
preshed==3.0.8
protobuf==4.22.1
pydantic==1.10.7
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
sentencepiece==0.1.97
smart-open==6.3.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
thinc==8.1.9
tokenizers==0.13.3
torch==2.0.0
tqdm==4.65.0
transformers==4.27.4
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
